Support bazel builds
