+++ BUILD.bazel
@@ -1,0 +1,217 @@
+load("@dpdk_pip//:requirements.bzl", "requirement")
+load("@rules_foreign_cc//foreign_cc:meson.bzl", "meson_with_requirements")
+
+filegroup(
+    name = "srcs",
+    srcs = glob(["**"]),
+    visibility = ["//visibility:private"],
+)
+
+meson_with_requirements(
+    name = "dpdk",
+    lib_source = ":srcs",
+    out_headers_only = True,
+    out_static_libs = [
+        "librte_acl.a",
+        "librte_dispatcher.a",
+        "librte_metrics.a",
+        "librte_net_tap.a",
+        "librte_argparse.a",
+        "librte_distributor.a",
+        "librte_ml_cnxk.a",
+        "librte_net_thunderx.a",
+        "librte_baseband_acc.a",
+        "librte_dma_cnxk.a",
+        "librte_mldev.a",
+        "librte_net_txgbe.a",
+        "librte_baseband_fpga_5gnr_fec.a",
+        "librte_dmadev.a",
+        "librte_net.a",
+        "librte_net_vdev_netvsc.a",
+        "librte_baseband_fpga_lte_fec.a",
+        "librte_dma_dpaa2.a",
+        "librte_net_af_packet.a",
+        "librte_net_vhost.a",
+        "librte_baseband_la12xx.a",
+        "librte_dma_dpaa.a",
+        "librte_net_ark.a",
+        "librte_net_virtio.a",
+        "librte_baseband_null.a",
+        "librte_dma_hisilicon.a",
+        "librte_net_atlantic.a",
+        "librte_net_vmxnet3.a",
+        "librte_baseband_turbo_sw.a",
+        "librte_dma_idxd.a",
+        "librte_net_avp.a",
+        "librte_net_xsc.a",
+        "librte_bbdev.a",
+        "librte_dma_ioat.a",
+        "librte_net_axgbe.a",
+        "librte_net_zxdh.a",
+        "librte_bitratestats.a",
+        "librte_dma_odm.a",
+        "librte_net_bnxt.a",
+        "librte_node.a",
+        "librte_bpf.a",
+        "librte_dma_skeleton.a",
+        "librte_net_bond.a",
+        "librte_pcapng.a",
+        "librte_bus_auxiliary.a",
+        "librte_eal.a",
+        "librte_net_cnxk.a",
+        "librte_pci.a",
+        "librte_bus_cdx.a",
+        "librte_efd.a",
+        "librte_net_cpfl.a",
+        "librte_pdcp.a",
+        "librte_bus_dpaa.a",
+        "librte_ethdev.a",
+        "librte_net_cxgbe.a",
+        "librte_pdump.a",
+        "librte_bus_fslmc.a",
+        "librte_event_cnxk.a",
+        "librte_net_dpaa2.a",
+        "librte_pipeline.a",
+        "librte_bus_ifpga.a",
+        "librte_eventdev.a",
+        "librte_net_dpaa.a",
+        "librte_pmu.a",
+        "librte_bus_pci.a",
+        "librte_event_dlb2.a",
+        "librte_net_e1000.a",
+        "librte_port.a",
+        "librte_bus_platform.a",
+        "librte_event_dpaa2.a",
+        "librte_net_ena.a",
+        "librte_power.a",
+        "librte_bus_uacce.a",
+        "librte_event_dpaa.a",
+        "librte_net_enetc.a",
+        "librte_power_acpi.a",
+        "librte_bus_vdev.a",
+        "librte_event_dsw.a",
+        "librte_net_enetfec.a",
+        "librte_power_amd_pstate.a",
+        "librte_bus_vmbus.a",
+        "librte_event_octeontx.a",
+        "librte_net_enic.a",
+        "librte_power_cppc.a",
+        "librte_cfgfile.a",
+        "librte_event_opdl.a",
+        "librte_net_failsafe.a",
+        "librte_power_intel_pstate.a",
+        "librte_cmdline.a",
+        "librte_event_skeleton.a",
+        "librte_net_fm10k.a",
+        "librte_power_intel_uncore.a",
+        "librte_common_cnxk.a",
+        "librte_event_sw.a",
+        "librte_net_gve.a",
+        "librte_power_kvm_vm.a",
+        "librte_common_cpt.a",
+        "librte_fib.a",
+        "librte_net_hinic3.a",
+        "librte_raw_cnxk_bphy.a",
+        "librte_common_dpaax.a",
+        "librte_gpudev.a",
+        "librte_net_hinic.a",
+        "librte_raw_cnxk_gpio.a",
+        "librte_common_ionic.a",
+        "librte_graph.a",
+        "librte_net_hns3.a",
+        "librte_raw_cnxk_rvu_lf.a",
+        "librte_common_nfp.a",
+        "librte_gro.a",
+        "librte_net_i40e.a",
+        "librte_rawdev.a",
+        "librte_common_nitrox.a",
+        "librte_gso.a",
+        "librte_net_iavf.a",
+        "librte_raw_dpaa2_cmdif.a",
+        "librte_common_octeontx.a",
+        "librte_hash.a",
+        "librte_net_ice.a",
+        "librte_raw_gdtc.a",
+        "librte_common_qat.a",
+        "librte_ip_frag.a",
+        "librte_net_idpf.a",
+        "librte_raw_ntb.a",
+        "librte_common_sfc_efx.a",
+        "librte_ipsec.a",
+        "librte_net_ionic.a",
+        "librte_raw_skeleton.a",
+        "librte_common_zsda.a",
+        "librte_jobstats.a",
+        "librte_net_ixgbe.a",
+        "librte_rcu.a",
+        "librte_compressdev.a",
+        "librte_kvargs.a",
+        "librte_net_memif.a",
+        "librte_regex_cn9k.a",
+        "librte_compress_nitrox.a",
+        "librte_latencystats.a",
+        "librte_net_netvsc.a",
+        "librte_regexdev.a",
+        "librte_compress_octeontx.a",
+        "librte_log.a",
+        "librte_net_nfp.a",
+        "librte_reorder.a",
+        "librte_crypto_bcmfs.a",
+        "librte_lpm.a",
+        "librte_net_ngbe.a",
+        "librte_rib.a",
+        "librte_crypto_caam_jr.a",
+        "librte_mbuf.a",
+        "librte_net_ntnic.a",
+        "librte_ring.a",
+        "librte_crypto_cnxk.a",
+        "librte_member.a",
+        "librte_net_null.a",
+        "librte_sched.a",
+        "librte_cryptodev.a",
+        "librte_mempool.a",
+        "librte_net_octeon_ep.a",
+        "librte_security.a",
+        "librte_crypto_dpaa2_sec.a",
+        "librte_mempool_bucket.a",
+        "librte_net_octeontx.a",
+        "librte_stack.a",
+        "librte_crypto_dpaa_sec.a",
+        "librte_mempool_cnxk.a",
+        "librte_net_pfe.a",
+        "librte_table.a",
+        "librte_crypto_ionic.a",
+        "librte_mempool_dpaa2.a",
+        "librte_net_qede.a",
+        "librte_telemetry.a",
+        "librte_crypto_nitrox.a",
+        "librte_mempool_dpaa.a",
+        "librte_net_r8169.a",
+        "librte_timer.a",
+        "librte_crypto_null.a",
+        "librte_mempool_octeontx.a",
+        "librte_net_ring.a",
+        "librte_vdpa_ifc.a",
+        "librte_crypto_octeontx.a",
+        "librte_mempool_ring.a",
+        "librte_net_rnp.a",
+        "librte_vdpa_nfp.a",
+        "librte_crypto_scheduler.a",
+        "librte_mempool_stack.a",
+        "librte_net_sfc.a",
+        "librte_vdpa_sfc.a",
+        "librte_crypto_virtio.a",
+        "librte_meter.a",
+        "librte_net_softnic.a",
+        "librte_vhost.a",
+    ],
+    postfix_script = "ninja",
+    requirements = [
+        requirement("pyelftools"),
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "@libnuma",
+        "@openssl//:crypto",
+    ],
+)
