+++ MODULE.bazel
@@ -1,0 +1,28 @@
+module(
+    name = "dpdk",
+    repo_name = "dpdk",
+)
+
+bazel_dep(name = "rules_foreign_cc", version = "0.15.1")
+bazel_dep(name = "rules_python", version = "1.6.3")
+bazel_dep(name = "openssl", version = "3.3.1.bcr.7")
+
+tools = use_extension("@rules_foreign_cc//foreign_cc:extensions.bzl", "tools")
+use_repo(tools, "meson_src")
+
+python = use_extension("@rules_python//python/extensions:python.bzl", "python")
+python.toolchain(
+    python_version = "3.12",
+)
+use_repo(python, "python_versions")
+
+pip = use_extension("@rules_python//python/extensions:pip.bzl", "pip")
+pip.parse(
+    hub_name = "dpdk_pip",
+    python_version = "3.12",
+    requirements_lock = "//bazel:requirements.txt",
+)
+use_repo(pip, "dpdk_pip")
+
+cpp_deps = use_extension("//bazel:deps.bzl", "cpp_deps_extension")
+use_repo(cpp_deps, "libnuma")
