+++ bazel/requirements.in
@@ -1,0 +1,1 @@
+pyelftools==0.32
